@@ -15,61 +15,63 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/core": "7.1.4",
-    "@angular/animations": "7.1.4",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "7.1.4",
-    "@angular/compiler": "7.1.4",
-    "@angular/core": "7.1.4",
-    "@angular/flex-layout": "^7.0.0-beta.24",
-    "@angular/forms": "7.1.4",
-    "@angular/http": "7.1.4",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "7.1.4",
-    "@angular/platform-browser-dynamic": "7.1.4",
-    "@angular/router": "7.1.4",
-    "@iplab/ngx-file-upload": "^1.4.0",
-    "braces": "^2.3.2",
-    "core-js": "^2.6.9",
+    "@angular-devkit/core": "8.3.20",
+    "@angular/animations": "8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "8.2.14",
+    "@angular/compiler": "8.2.14",
+    "@angular/core": "8.2.14",
+    "@angular/flex-layout": "^8.0.0-beta.27",
+    "@angular/forms": "8.2.14",
+    "@angular/http": "8.0.0-beta.10",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "8.2.14",
+    "@angular/platform-browser-dynamic": "8.2.14",
+    "@angular/router": "8.2.14",
+    "@iplab/ngx-file-upload": "^1.4.1",
+    "braces": "^3.0.2",
+    "core-js": "^3.5.0",
     "diff": "^4.0.1",
     "file-saver": "^2.0.2",
     "font-awesome": "^4.7.0",
     "fstream": "^1.0.12",
     "ngx-gallery": "^5.10.0",
-    "ngx-page-scroll": "^5.0.1",
+    "ngx-page-scroll": "^7.0.0",
+    "ngx-page-scroll-core": "^7.0.0",
     "npm": "^6.13.4",
-    "primeng": "^7.1.3",
-    "rxjs": "^6.5.2",
-    "snyk": "^1.192.4",
-    "sshpk": "^1.13.2",
-    "tar": "^2.2.2",
-    "tar-fs": "^1.16.3",
+    "primeng": "^9.0.0-rc.2",
+    "rxjs": "^6.5.3",
+    "snyk": "^1.260.0",
+    "sshpk": "^1.16.1",
+    "tar": "^5.0.5",
+    "tar-fs": "^2.0.0",
     "tslib": "^1.10.0",
     "web-animations-js": "^2.3.2",
-    "zone.js": "^0.8.29"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.9",
-    "@angular/cli": "7.1.4",
-    "@angular/compiler-cli": "8.0.0",
-    "@angular/language-service": "7.1.4",
-    "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "^2.0.6",
-    "@types/node": "^6.14.7",
-    "angular-ide": "^0.9.65",
-    "codelyzer": "^4.4.0",
-    "jasmine-core": "~3.3.0",
+    "@angular-devkit/build-angular": "^0.803.20",
+    "@angular/cli": "8.3.10",
+    "@angular/compiler-cli": "8.2.14",
+    "@angular/language-service": "8.2.14",
+    "@types/jasmine": "~3.5.0",
+    "@types/jasminewd2": "^2.0.8",
+    "@types/node": "^12.12.17",
+    "angular-ide": "^0.9.70",
+    "codelyzer": "^5.2.0",
+    "eslint":   "^6.7.2",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.0.1",
-    "karma-chrome-launcher": "~2.2.0",
+    "karma": "~4.4.1",
+    "karma-chrome-launcher": "~3.1.0",
     "karma-cli": "~2.0.0",
-    "karma-coverage-istanbul-reporter": "^2.1.0",
+    "karma-coverage-istanbul-reporter": "^2.1.1",
     "karma-jasmine": "^2.0.1",
     "karma-jasmine-html-reporter": "^1.4.2",
     "protractor": "^6.0.0",
-    "ts-node": "~7.0.1",
-    "tslint": "~5.8.0",
-    "typescript": "~3.4.5"
+    "ts-node": "~8.5.4",
+    "tslint": "~5.20.1",
+    "typescript": "~3.5.3"
   },
   "snyk": true
 }
