@@ -5,7 +5,7 @@ import { AppModule } from './app/app.module';
 import { environment } from './environments/environment';
 
 if (environment.production) {
-  enableProdMode();
+//  enableProdMode();
 }
 
 platformBrowserDynamic().bootstrapModule(AppModule)
